Found new land.
A new land has been found! Another island, smaller than my own home base was the first place that I saw, no more than 200b (blocks) from my dock.  I constructed a dock at the new land, then disembarked to look around.

I found that the land was populated with sheep, and chickens, a pleasant sight, as animals were not to be found at my home.  As well, new birch trees were growing -- their light wood will be lovely in my home in contrast to the oakwood, and their saplings will grow wonderfully in my forest.

I've built a roadway across the sea from my home island to here.  It was short work, but building safe landing areas for the railway I plan to build was dangerous work, underwater.  I also built a stronghold on the new land.  There are more monsters here, I must note, as the land is not lit by torches as mine is.

I discovered that this was but the tip of the land mass ahead of me.  I am not sure how far it goes, but it carries on as far as I can see.  It may yet be another, larger island, or the start of a grand continent.  I intend to find out.